requirements.txt
requests 
streamlit 
Pillow